Main (lines 25-34)
This section of the code analyzes the user input by storing it in a temporary variable and checking if it is less than 25, equal to zero, or less than zero. If any of these go against the stipulated restriction then it is flagged as illegal. If it is legal, the code goes on to generate the Fibonacci sequence.