Fibonacci Loop (lines 78-86)
This section of the code decrements the counter to track how many values have been added to the sequence and checks if more values must be printed. If the illegal check called in the main section is done and no more numbers need to be added, then the loop and the program ends.